// frontend/src/types/category.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Category = { id: number, name: string, parent_id: number | null, image_id: string | null, position: number, is_active: boolean, created_by: number, created_at: string, updated_at: string, };

export type NewCategory = { name: string, parent_id?: number, image_id?: string, };

export type UpdateCategory = { name?: string, parent_id?: number | null, image_id?: string | null, position?: number, };